move plantuml namespace into domain
